Moved polynomial functionality to polynomit.c also created polynomit.h
