+++ aaro_larange2.vcxproj
@@ -140,6 +140,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="main.c" />
+    <ClCompile Include="polynomit.c" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="polynomit.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
